+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Unreleased
 * Add support for `Room Resource`  #307
 * Fix issue where "302 Redirect" response codes were treated as errors #306
+* Fix issue where `api` value was overriden when calling `message.expanded` #311
 
 ### 5.1.0 / 2021-06-07
 * Add support for read only attributes #298